% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{candidate_ffls}
\alias{candidate_ffls}
\title{candidate_ffls}
\usage{
candidate_ffls(mirna_expr, mrna_expr)
}
\arguments{
\item{mirna_expr}{Dataframe (miRNA x samples) of miRNA expression data with miRNAs in accession number format}

\item{mrna_expr}{Dataframe (mRNA x samples) of mRNA expression data with genes in EnsemblID format}
}
\value{
Vector of two dataframes (miRNA-FFLs and TF-FFLs)
}
\description{
Generate candidate miRNA-FFLs and TF-FFLs
}
