% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{candidate_ffls_mirna_all_part2}
\alias{candidate_ffls_mirna_all_part2}
\title{candidate_ffls_mirna_all_part2}
\format{A DataFrame (13240974 x 9) with the following columns:
\describe{
\item{mirna}{miRNA in candidate FFL [Accession Number]}
\item{tf}{TF in candidate FFL [Ensembl ID]}
\item{targetgene}{target gene in candidate FFL [Ensembl ID]}
\item{TARGETSCAN}{indicator of whether miRNA-target gene or miRNA-TF interaction is in TargetScan v7.2 database}
\item{MIRTARBASE}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRTarBase v7.0 database}
\item{MIRDB}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRDB v6.0 database}
\item{MIRANDA}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRanda v3.0 database}
\item{TRRUST}{indicator of whether TF-target gene or miRNA-TF interaction is in TRRUST v2.0 database}
\item{ENCODE}{indicator of whether TF-target gene or miRNA-TF interaction is in ENCODE database}
}}
\usage{
candidate_ffls_mirna_all_part2
}
\description{
Second half of all possible candidate miRNA FFLs assembled from databases
}
\keyword{datasets}
