% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RM_hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(x)
}
\arguments{
\item{x}{The name of the person to address}
}
\value{
The output from \code{\link{print}}
}
\description{
Print a hello statement
}
\examples{
hello("John")
}
