% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{mirna_targetgene_db}
\alias{mirna_targetgene_db}
\title{miRNA-Target Gene Interactions}
\format{A DataFrame (2477728 x 6) with the following columns:
\describe{
\item{mirna}{miRNA involved in the interaction [Accession Number]}
\item{targetgene}{target gene involved in the interaction (gene targeted by miRNA) [Ensembl ID]}
\item{TARGETSCAN}{indicator of whether interaction is in TargetScan v7.2 database}
\item{MIRTARBASE}{indicator of whether interaction is in miRTarBase v7.0 database}
\item{MIRDB}{indicator of whether interaction is in miRDB v6.0 database}
\item{MIRANDA}{indicator of whether interaction is in miRanda v3.0 database}
}}
\usage{
mirna_targetgene_db
}
\description{
A DataFrame of miRNA-target gene interactions
compiled from databases (TargetScan v7.2, miRTarBase v7.0,
miRDB v6.0, and miRanda v3.0).
}
\keyword{datasets}
