% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{names_mirna_db}
\alias{names_mirna_db}
\title{miRNAs}
\format{
Vector of miRNAs (4543 miRNAs)
}
\usage{
names_mirna_db
}
\description{
miRNAs covered among all databases
}
\keyword{datasets}
