% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{names_targetgene_db}
\alias{names_targetgene_db}
\title{Target Genes}
\format{
Vector of target genes (22183 target genes)
}
\usage{
names_targetgene_db
}
\description{
Target genes covered among all databases
}
\keyword{datasets}
