% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{names_tf_db}
\alias{names_tf_db}
\title{Transcription Factors}
\format{Vector of transcription factors (1149 transcription factors)}
\usage{
names_tf_db
}
\description{
Transcription factors covered among all databases
}
\keyword{datasets}
