% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_predict_ffls_two_groups.R
\name{predict_ffls_two_groups}
\alias{predict_ffls_two_groups}
\title{predict_ffls_two_groups}
\usage{
predict_ffls_two_groups(mirna_expr_g1, mrna_expr_g1, mirna_expr_g2,
  mrna_expr_g2, ffl_type = c("miRNA", "TF"), candidate_ffls,
  first_row = 1, last_row = nrow(candidate_ffls),
  num_bootstrap_samples = 1000, num_permutations = 1000,
  p_value_adjust_method = c("holm", "hochberg", "hommel", "bonferroni",
  "BH", "BY", "fdr", "none"), seed = 12345)
}
\arguments{
\item{mirna_expr_g1}{Group 1's miRNA expression data (dataframe: miRNAs x samples)}

\item{mrna_expr_g1}{Group 1's mRNA expression data (dataframe: mRNAs x samples)}

\item{mirna_expr_g2}{Group 2's miRNA expression data (dataframe: miRNAs x samples, see sample_mirna_expr)}

\item{mrna_expr_g2}{Group 2's miRNA expression data (dataframe: miRNAs x samples, see sample_mirna_expr)}

\item{ffl_type}{FFL type (character: "miRNA" or "TF")}

\item{candidate_ffls}{candidate FFLs (dataframe: candidate FFLs x 10, see sample_candidate_ffls)}

\item{first_row}{first row of candidate FFL dataframe to include in analyses (integer: default = 1)}

\item{last_row}{last row of candidate FFL dataframe to include in analyses (integer: default = nrow(candidate_ffls))}

\item{num_bootstrap_samples}{number of bootstrap samples (integer: default = 1000)}

\item{num_permutations}{number of permutations (integer: default = 1000)}

\item{p_value_adjust_method}{method to adjust p-values for multiple testing (character: "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", or "none")}

\item{seed}{random seed (integer: default = 12345)}
}
\value{
candidate FFLs with delta-pFFLs and p-values (dataframe: candidate FFLs x 18, see sample output)
}
\description{
function to predict FFLs unique to one of two biological groups
}
