% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prll_functions.R
\name{predict_mirna_ffls}
\alias{predict_mirna_ffls}
\title{predict_ffls}
\usage{
predict_mirna_ffls(mirna_expr, mrna_expr, ffl_type, num_bootstrap_samples,
  alpha_mediation, num_permutations, alpha_permutation_test, seed)
}
\arguments{
\item{mirna_expr}{Dataframe (miRNA x samples) of miRNA expression data with miRNAs in accession number format}

\item{mrna_expr}{Dataframe (mRNA x samples) of mRNA expression data with genes in EnsemblID format}

\item{ffl_type}{Character ("miRNA" or "TF") indicating the FFL type (miRNA-FFL or TF-FFL)}

\item{num_bootstrap_samples}{Number of bootstrap samples when calculating p(FFL) (default is 1000)}

\item{alpha_mediation}{Significance level of coefficients in the mediation model's linear equations (default is 0.05)}

\item{num_permutations}{Number of permutations in permutation test (default is 1000)}

\item{alpha_permutation_test}{Significance level of FFLs from permutation test (default is 0.05)}

\item{seed}{random seed for bootstrapping and permutation test (default is 12345)}
}
\value{
Dataframes of candidate FFLs that meet mediation model conditions and are statistically significant)
}
\description{
Predict FFLs using mediation model approach
}
