% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_candidate_ffls_mirna}
\alias{sample_candidate_ffls_mirna}
\title{sample_candidate_ffls_mirna}
\format{
A DataFrame (3 x 9) with the following columns:
\describe{
\item{mirna}{miRNA in candidate FFL [Accession Number]}
\item{tf}{TF in candidate FFL [Ensembl ID]}
\item{targetgene}{target gene in candidate FFL [Ensembl ID]}
\item{TARGETSCAN}{indicator of whether miRNA-target gene or miRNA-TF interaction is in TargetScan v7.2 database}
\item{MIRTARBASE}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRTarBase v7.0 database}
\item{MIRDB}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRDB v6.0 database}
\item{MIRANDA}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRanda v3.0 database}
\item{TRRUST}{indicator of whether TF-target gene or miRNA-TF interaction is in TRRUST v2.0 database}
\item{ENCODE}{indicator of whether TF-target gene or miRNA-TF interaction is in ENCODE database}
}
}
\usage{
sample_candidate_ffls_mirna
}
\description{
Sample candidate miRNA FFLs serving as input for \code{\link{predict_ffls_one_group}} or \code{\link{predict_ffls_two_groups}}
}
\keyword{datasets}
