% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_mirna_expr}
\alias{sample_mirna_expr}
\title{sample_mirna_expr}
\format{A DataFrame (3 x 50) representing expression data of 3 miRNAs for 50 samples}
\usage{
sample_mirna_expr
}
\description{
Sample mirna_expr serving as input for \code{\link{predict_ffls_one_group}}
}
\keyword{datasets}
