% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_mirna_expr_g2}
\alias{sample_mirna_expr_g2}
\title{sample_mirna_expr_g2}
\format{
A DataFrame (3 x 50) representing expression data of 3 miRNAs for 50 samples in Group 2
}
\usage{
sample_mirna_expr_g2
}
\description{
Sample mirna_expr_g2 serving as input for \code{\link{predict_ffls_two_groups}}
}
\keyword{datasets}
