% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_mrna_expr}
\alias{sample_mrna_expr}
\title{sample_mrna_expr}
\format{
A DataFrame (5 x 50) representing expression data of 5 mRNAs for 50 samples
}
\usage{
sample_mrna_expr
}
\description{
Sample mrna_expr serving as input for \code{\link{predict_ffls_one_group}}
}
\keyword{datasets}
