% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_mrna_expr_g1}
\alias{sample_mrna_expr_g1}
\title{sample_mrna_expr_g1}
\format{
A DataFrame (5 x 50) representing expression data of 5 mRNAs for 50 samples in Group 1
}
\usage{
sample_mrna_expr_g1
}
\description{
Sample mrna_expr_g1 serving as input for \code{\link{predict_ffls_two_groups}}
}
\keyword{datasets}
