% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_info.R
\docType{data}
\name{sample_output_twogroups_mirna_ffls}
\alias{sample_output_twogroups_mirna_ffls}
\title{sample_output_twogroups_mirna_ffls}
\format{A DataFrame (3 x 85) representing expression data of 5 mRNAs for 50 samples
\describe{
\item{mirna_name}{miRNA in candidate FFL [miRNA name]}
\item{tf_name}{TF in candidate FFL [SYMBOL]}
\item{targetgene_name}{target gene in candidate FFL [SYMBOL]}
\item{pFFL_group1}{pFFL of candidate FFL in Group 1}
\item{pFFL_group2}{pFFL of candidate FFL in Group 2}
\item{delta_pFFL_(group1-group2)}{delta-pFFL of candidate FFL}
\item{p_value_adj}{adjusted p-value of candidate FFL}
\item{p_value}{unadjusted p-value of candidate FFL}
\item{mirna}{miRNA in candidate FFL [Accession Number]}
\item{tf}{TF in candidate FFL [Ensembl ID]}
\item{targetgene}{target gene in candidate FFL [Ensembl ID]}
\item{TARGETSCAN}{indicator of whether miRNA-target gene or miRNA-TF interaction is in TargetScan v7.2 database}
\item{MIRTARBASE}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRTarBase v7.0 database}
\item{MIRDB}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRDB v6.0 database}
\item{MIRANDA}{indicator of whether miRNA-target gene or miRNA-TF interaction is in miRanda v3.0 database}
\item{TRRUST}{indicator of whether TF-target gene or miRNA-TF interaction is in TRRUST v2.0 database}
\item{ENCODE}{indicator of whether TF-target gene or miRNA-TF interaction is in ENCODE database}
\item{remaining columns}{estimate, standard error (se), t-statistic (t), and p-value of regression coefficients for group1 and group2}
}}
\usage{
sample_output_twogroups_mirna_ffls
}
\description{
Sample output of \code{\link{predict_ffls_two_groups}}
}
\keyword{datasets}
