% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{step2_mediation}
\alias{step2_mediation}
\title{step2_mediation}
\usage{
step2_mediation(mirna_expr, mrna_expr, candidate_ffls,
  ffl_type = c("miRNA", "TF"), alpha = 0.05)
}
\arguments{
\item{mirna_expr}{Dataframe (miRNA x samples) of miRNA expression data with miRNAs in accession number format}

\item{mrna_expr}{Dataframe (mRNA x samples) of mRNA expression data with genes in EnsemblID format}

\item{candidate_ffls}{Dataframe of candidate ffls (output from \code{\link{step1_candidate_ffls}})}

\item{ffl_type}{Character ("miRNA" or "TF") indicating the FFL type (miRNA-FFL or TF-FFL)}

\item{alpha}{Significance level of coefficients in the mediation model's linear equations (default is 0.05)}
}
\value{
Dataframe of FFLs among candidate FFLs that meet the mediation model's requirements
}
\description{
Apply mediation model to each candidate FFL
}
