% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_functions.R
\name{step4_permutation_test}
\alias{step4_permutation_test}
\title{step4_permutation_test}
\usage{
step4_permutation_test(mirna_expr, mrna_expr, ffls, ffl_type = c("miRNA",
  "TF"), num_permutations, num_bootstrap_samples, alpha, seed)
}
\arguments{
\item{mirna_expr}{Dataframe (miRNA x samples) of miRNA expression data with miRNAs in accession number format}

\item{mrna_expr}{Dataframe (mRNA x samples) of mRNA expression data with genes in EnsemblID forma}

\item{ffls}{Dataframe of FFLs that meet mediation model conditions (output of step2_mediation)}

\item{ffl_type}{Character ("miRNA" or "TF") indicating the FFL type (miRNA-FFL or TF-FFL}

\item{num_permutations}{Number of permutations in permutation test (default is 1000)}

\item{num_bootstrap_samples}{Number of bootstrap samples when calculating p(FFL) (default is 1000)}

\item{alpha}{Significance level of coefficients in the mediation model's linear equations}

\item{seed}{Random seed}
}
\value{
\code{ffls} dataframe with added column of p-values associated with p(FFL) values
}
\description{
Calculate p-value associated with p(FFL) for FFLs that meet mediation model conditions through a permutation test
}
