% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{tf_mirna_db}
\alias{tf_mirna_db}
\title{Transcription Factor (TF)-miRNA Interactions}
\format{A DataFrame (117543 x 3) with the following columns:
\describe{
\item{tf}{TF involved in the interaction}
\item{mirna}{miRNA involved in the interaction (miRNA targeted by TF)}
\item{TRANSMIR}{indicator of whether interaction is in TransmiR v2.0 database}
}}
\usage{
tf_mirna_db
}
\description{
A DataFrame of TF-miRNA interactions
compiled from databases (TransmiR v2.0).
}
\keyword{datasets}
