% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_info.R
\docType{data}
\name{tf_targetgene_db}
\alias{tf_targetgene_db}
\title{Transcription Factor (TF)-Target Gene Interactions}
\format{A DataFrame (1624664 x 4) with the following columns:
\describe{
\item{tf}{TF involved in the interaction [Ensembl ID]}
\item{targetgene}{target gene involved in the interaction (gene targeted by TF), target genes include transcription factors [Ensembl ID]}
\item{TRRUST}{indicator of whether interaction is in TRRUST v2.0 database}
\item{ENCODE}{indicator of whether interaction is in ENCODE database}
}}
\usage{
tf_targetgene_db
}
\description{
A DataFrame of TF-target gene interactions
compiled from databases (ENCODE and TRRUST v2.0).
}
\keyword{datasets}
