% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prll_functions.R
\name{mediation_ffl}
\alias{mediation_ffl}
\title{mediation_ffl}
\usage{
mediation_ffl(mirna, tf, targetgene, ffl_type = c("miRNA", "TF"), alpha)
}
\arguments{
\item{mirna}{Vector of miRNA expression data}

\item{tf}{Vector of mRNA expression data}

\item{targetgene}{Vector target gene expression data}

\item{ffl_type}{Character ("miRNA" or "TF") indicating the FFL type (miRNA-FFL or TF-FFL)}

\item{alpha}{Significance level of coefficients in the mediation model's linear equations}
}
\value{
Boolean indicating whether the set of miRNA, TF, and target gene expression data meet the conditions of the mediation model
}
\description{
Determines whether a set of miRNA, TF, and target gene expression data meet the conditions of the mediation model
}
