% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_function.R
\name{predict_ffls}
\alias{predict_ffls}
\title{predict_ffls}
\usage{
predict_ffls(mirna_expr, mrna_expr, mediation_alpha = 0.05)
}
\arguments{
\item{mirna_expr}{Dataframe (miRNA x samples) of miRNA expression data with miRNAs in accession number format}

\item{mrna_expr}{Dataframe (mRNA x samples) of mRNA expression data with genes in EnsemblID format}

\item{mediation_alpha}{Significance level of coefficients in the mediation model's linear equations (default is 0.05)}
}
\value{
!!!!!not sure yet
}
\description{
Predict FFLs using mediation model approach
}
